requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
Click==7.0
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
Flask==1.1.1
Flask-PyMongo==2.3.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.1
jsonify==0.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Pygments==2.5.2
pymongo==3.10.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
wincertstore==0.2
zipp==2.2.0
